Merge ObjWriter into dotnet/main and update to LLVM 14 API (#287)
* Apply llvm.patch

Taken from https://github.com/dotnet/runtime/blob/7ab969c84ef05ba948c0075392716ce335b47744/src/coreclr/tools/aot/ObjWriter/llvm.patch.

* Add objwriter library

* Taken from https://github.com/dotnet/runtime/tree/7ab969c84ef05ba948c0075392716ce335b47744/src/coreclr/tools/aot/ObjWriter.
* Updated README.md
* Updated CMakeLists.txt to remove reference to CORECLR_INCLUDE_DIR.
* Added cordebuginfo.h, cvconst.h, cfi.h from coreclr/inc at the above commit.

* Build the ObjWriter package

* Add ObjWriter API to set DWARF version (#161)

Contributes to https://github.com/dotnet/runtimelab/issues/1738.

* Add `.note.GNU-stack` section to produced executables (#162)

Do this unconditionally because there's no scenario where we would need executable stack for managed code.

* Remove Darwin workaround (#163)

This caught my attention as I was looking at the ObjWriter. LLVM no longer emits a `LC_VERSION_MIN_MACOSX` load command unless we explicitly set a version. I don't see a difference in `llvm-objdump -macho -x foo.o` with/without these lines (I didn't bother myself to boot into macOS to run `otool`).

* Fix llvm-dwarfdump warnings (#164)

Fixes https://github.com/dotnet/runtimelab/issues/1535. No warnings left with llvm-dwarfdump from LLVM 12.

* Revert "Fix llvm-dwarfdump warnings (#164)" (#218)

This reverts commit afc9070f64d110ce2bf029a4a624b9dd1c6dbffd.

* Add new NuGet package, `Microsoft.NETCore.Runtime.JIT.Tools`, includes `FileCheck` and `llvm-mca` (#256)

https://github.com/dotnet/runtime is wanting to start writing assembly (x64/ARM64) verification tests. Instead of building our own tool to support writing those kinds of tests, we want to leverage LLVM's `FileCheck`.

We also want to include `llvm-mca` at the request of @EgorBo

This PR creates a new NuGet package for `dotnet/runtime` to consume which we named `Microsoft.NETCore.Runtime.JIT.Tools`. So far, this package only includes LLVM's `FileCheck` and `llvm-mca` tools.

* [ObjWriter] Enable DWARF debug information emitting for Mach-O (#269)

* Account for GOT VariantKind on osx-arm64 (#185)

* Add API for emitting compact unwind encoding, enforce DWARF encoding if not explicitly overridden

* Add comment

* Update ObjWriter to LLVM 14 API

* Add support for generating uninitialized sections (#306)

We support `.bss` but not custom sections that are bss-like. This adds such support.

* Do not indiscriminately create text section (#312)

If we ended up with nothing in the text section, this line would error LLVM out in:

https://github.com/dotnet/llvm-project/blob/3db8d68195c17386557f1a258312bbae4051dc05/llvm/lib/MC/ELFObjectWriter.cpp#L1458-L1459

Because we generate a reference to the empty text section in the `aranges` section.

I double checked and debugging on Linux still works fine without this. `SetCodeSectionAttribute` is an objwriter API and we have access to it from the managed side. We should be calling it from there if it's needed for something that I didn't realize (we do call it from the managed side for the `.managed` section, but that one actually has debug information generated, unlike `.text`).

* Fix off-by-one error in DWARF reg-reg location (#317)

The DWARF specification states that the form of an exprloc consists
of an unsigned LEB128 length value, followed by the encoded location bytes
of the specified length. For some reason we were adding one to the length
value being emitted. This looks incorrect to me. The above calculation for
REG-REG (a variable stored in two registers) correctly calculates the length
of each register type tag, plus the size of the interpolating PIECE tags,
plus the size of notation for each register. The extra byte looks wrong.

I've tested this locally and it appears to resolve https://github.com/dotnet/runtime/issues/77407.

Unfortunately, it also causes llvm-dwarfdump --verify to constantly
complain about missing base addresses. I can't confirm at the moment,
but my suspicion is that this is revealing an existing bug. Even if this
is somehow causing a new bug, I think the resulting symbols with this
change are better than the alternative (no working symbols at all).

* Setting context object file info

* Add verbosity to linux x64 pipeline

In order to understand what is happening with std path error.

* Revert "Add verbosity to linux x64 pipeline"

This reverts commit 5c4636e86b6ae027fc54646ba513c18a02bce642.

* Upgrading linux build image

* [Temporary] Adding verbosity to get more pipeline error info

* Update image name for linux x64

* Fix Linux x64 build

* Revert "[Temporary] Adding verbosity to get more pipeline error info"

This reverts commit 9d76b3675bb48c276045bb28239a177df54cc939.

* Updating Build_Linux_musl timeout

* Update linux-musl Docker images

* Fix linux-musl-x64 build

* Setting clang/++ version 15 for linux musl

* Copying clang/clan++ vars to unix-like OS

* Fix cut & paste error

* Fix objcopy and strip path in cross-compilation

* Update azure-pipelines.yml

$(ClangVersion) $(ClangPlusVersion) weren't defined for OSX and should be defined for every Linux

* Bump timeout for Linux musl build

* Clean up .gitignore

* Consolidate Clang[Plus]Version into ClangVersionArg

* Move CLANG_TARGET from environment into build parameter
Always quote _BuildConfig on command line so empty value is not accidentally using next parameter as the value

* Update URL in cordebuginfo.h to point to dotnet/runtime

* Bump Windows build timeout to 210

* Fix a typo in compiler name

* Revert $(_BuildConfig) -> "$(_BuildConfig)" change

* Change ClangTarget to ClangTargetArg since apparently it gets propagated as environment variable into wrong steps

* Fix inadvertent change

* Bump timeout everywhere

---------

Co-authored-by: Michal Strehovský <MichalStrehovsky@users.noreply.github.com>
Co-authored-by: Andy Gocke <andy@commentout.net>
Co-authored-by: Will Smith <lol.tihan@gmail.com>
Co-authored-by: Adeel Mujahid <3840695+am11@users.noreply.github.com>
Co-authored-by: Brian Bohe <brianbohe@gmail.com>
Co-authored-by: Alexander Köplinger <alex.koeplinger@outlook.com>