Bring proper fix for warnings about missing children (#369)
My previous change in #363 did not correctly account for all the possible items
that could create a child entry. This caused an extra null entry to be created,
which was treated as the end of the section.

This change is more thorough and features NoChildren abbrev sections, which should
hopefully be a little safer, as it will avoid having to produce extra null entries.
I've done some extra validation on this change to make sure that roughly the same amount
of DWARF info is produced in a test app before and after this change, so it's unlikely
that we're truncating most of the info again.